api/requirements.txt
flask
Pillow
tensorflow
numpy
scikit-image
opencv-python